day1/pyproject.toml
[project]
name = "day1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
